requirements.txt
google-cloud-bigquery==3.25.0
pandas==2.2.3
python-dotenv==1.0.1
google-genai~=1.46
fastapi~=0.110
uvicorn[standard]~=0.29
google-cloud-storage~=2.10.0